warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@10528 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/GuidedMissleStrategy.cxx
@@ -251,7 +251,7 @@ void GuidedMissileStrategy::update(float dt)
 float GuidedMissileStrategy::checkBuildings(const Ray& ray)
 {
   float shotSpeed = BZDB.eval(StateDatabase::BZDB_SHOTSPEED);
-  float t = (currentTime - prevTime) * shotSpeed;
+  float t = float((currentTime - prevTime) * shotSpeed);
   int face;
   const Obstacle* building = getFirstBuilding(ray, Epsilon, t);
   const Teleporter* teleporter = getFirstTeleporter(ray, Epsilon, t, face);
